+++ src/main/java/com/rensystem/apirest/apirest/dto/ProductDTO.java
@@ -7,6 +7,6 @@ public class ProductDTO {
     private Long id;
     private String nombre;
     private Double precio;
-    private Integer orden;
+    private Long orden;
 }
